feat: introduce RouteData and RouteInstructionItem types for improved route handling in MapController
