+++ roamance/src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './auth';
 export * from './api';
+export * from './map';
 export * from './places';
 export * from './navigation';
